backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
python-multipart==0.0.9
pydantic==2.7.1
pydantic-settings==2.2.1

# Ollama
ollama==0.2.1

# LangChain
langchain==0.2.1
langchain-community==0.2.1

# Document parsing
pymupdf==1.24.3
python-docx==1.1.2

# Vector store
faiss-cpu==1.8.0
numpy==1.26.4

# Utilities
python-dotenv==1.0.1
aiofiles==23.2.1
httpx==0.27.0

# Testing
pytest==8.2.0
pytest-asyncio==0.23.7